Add Delete User Profile Functionality and Authorization Middleware
This commit introduces the ability for users (or admins) to delete a user profile,
along with the necessary middleware to ensure proper authorization for this action.
The new `deleteUserProfileCtrl` function handles the deletion process, which includes
deleting the user's profile photo from Cloudinary and removing the user from the
database. Additionally, a new middleware function, `verifyTokenAndAuthorization`,
has been added to verify that the request is made by either the same user or an admin.
This middleware is utilized in the user routes to protect the delete profile endpoint.