Add Like/Unlike Feature for Posts
Implemented a new feature that allows users to like or unlike posts. This functionality is accessible via a PUT request to `/api/posts/like/:id`. The feature checks if the post is already liked by the logged-in user and toggles the like status accordingly. The changes include adding the `toggleLikeCtrl` controller method and a corresponding route in `postRoute.js`. This enhancement improves user engagement by providing a straightforward way to interact with posts.