Add Comments to Post Retrieval and Populate in Post Model
This commit enhances the post retrieval functionality by including comments associated with each post when fetching a single post. It modifies the `getSinglePostsCtrl` function in the `postController.js` to populate the `comments` field along with the existing user information. Additionally, it updates the `PostSchema` in the `Post.js` model to include virtual population of comments, enabling the association of comments with their respective posts. This change is aimed at improving the API's response by providing a more comprehensive view of posts, including their comments, directly from the post retrieval endpoint.