Add Functionality to Delete User Posts and Comments
When a user is deleted, their posts and comments should also be removed from the system. This commit introduces the necessary changes to ensure that when a user is deleted, all associated posts and comments are also deleted. This includes:

- Adding the `Comment` model to `postController.js` to allow for the deletion of comments associated with a post when the post is deleted.
- Implementing the deletion of all comments associated with a post in the `DeletePostsCtrl` function.
- Importing and using the `cloudinaryRemoveMultipleImg` function in `usersController.js` to delete all post images associated with a user from Cloudinary when the user is deleted.
- Adding the `cloudinaryRemoveMultipleImg` function to `utils/cloudinary.js` to handle the deletion of multiple images from Cloudinary.
- Updating the `deleteUserProfileCtrl` function to delete all posts and comments associated with a user when the user is deleted.

This ensures that all user-related data is properly cleaned up, maintaining data integrity and privacy.